requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
backcall==0.2.0
bcrypt==3.2.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.15
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.13.1
easysnmp==0.2.6
get-mac==0.8.3
idna==3.3
importlib-metadata==4.8.3
#ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
kombu==5.1.0
#nmap==0.0.1
nornir==2.4.0
ntlm-auth==1.5.0
packaging==21.3
paramiko==2.11.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
#pkg_resources==0.0.0
ply==3.11
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.9
pysmi==0.3.4
pysnmp==4.4.12
pysnmp-mibs==0.1.6
pytz==2022.2
pywinrm==0.4.3
redis==4.3.4
requests==2.27.1
requests-ntlm==1.1.0
scapy==2.4.5
six==1.16.0
spur==0.3.22
sqlparse==0.4.2
toml==0.10.2
traitlets==4.3.3
#typing_extensions==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
XlsxWriter==3.0.3
xmltodict==0.13.0
zipp==3.6.0
